chore: moving minio
